+++ NEWS.md
@@ -11,6 +11,9 @@
 * all: added new scalable (SVG) application icons for all applications,
   replacing the existing pixmap (PNG, ICO) icons. Icons from Sarreq
   Teryx. Implements #6253.
+* MKVToolNix GUI: job queue: added a context menu for opening copies of the
+  selected queue jobs in the multiplexer as new settings there without
+  removing the queue jobs. Implements #6255.
 
 ## Bug fixes